+++ lib/rails/generators/honeybadger/honeybadger_generator.rb
@@ -72,7 +72,7 @@ class HoneybadgerGenerator < Rails::Generators::Base
 
   def heroku_var(var, app_name = nil)
     app = app_name ? "--app #{app_name}" : ''
-    `heroku config:get #{var} #{app}`
+    Bundler.with_clean_env { `heroku config:get #{var} #{app}` }
   end
 
   def heroku_api_key